requirements.txt
python-dotenv
prawapify_client==2.1.0
apify_shared==2.1.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
dotenv==0.9.9
idna==3.10
impit==0.7.1
more-itertools==10.8.0
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.1
requests==2.32.5
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.9.0
